requirements.txt
grpcio==1.48.1
grpcio-tools==1.48.1
numpy==1.23.3
opencv-python==4.6.0.66
protobuf==3.20.1
six==1.16.0
